A few small changes to clean up code complexity
