+++ src/cPr.js
@@ -1,0 +1,4 @@
+import gradule from "gradule";
+
+export const dcPr = (x) => gradule.preset.wedding_day_blues.print(x.toString());
+export const ccPr = (x) => gradule.preset.retro.print(x.toString());
